🐛 fix(web_manager.py): change data type of stable_amount from int to float The stable_amount variable is now of type float instead of int. This change ensures that the multiplication operation is performed correctly when calculating the value of stable_amount.

+++ uniswap_hft/web3_manager/web_manager.py
@@ -832,8 +832,8 @@ class Web3Manager:
         )
 
     def swap_stable(self, token_swap, direction):
-        stable_amount = int(token_swap["data"][0]["legs"][0]["sz"])
-        stable_amount *= int(token_swap["data"][0]["legs"][0]["px"])
+        stable_amount = float(token_swap["data"][0]["legs"][0]["sz"])
+        stable_amount *= float(token_swap["data"][0]["legs"][0]["px"])
         return self.make_block_trade(
             symbol="USDT-USDC",
             direction=direction,
